requirements.txt
Flask
ngrok
line-bot-sdk
watchdog
openai
